+++ shinobi_client/orms/monitor.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional, List, Set
 import requests
 
 from shinobi_client.client import ShinobiClient
-from shinobi_client._common import raise_if_errors
+from shinobi_client._common import raise_if_errors, wait_and_verify
 
 
 @dataclass
@@ -49,6 +49,23 @@ class ShinobiMonitorOrm:
         :return:
         """
         return dict(filter(lambda item: item[0] in ShinobiMonitorOrm.SUPPORTED_KEYS, configuration.items()))
+
+    @staticmethod
+    def is_configuration_equivalent(configuration_1: Dict, configuration_2: Dict) -> bool:
+        """
+        TODO
+        :param configuration_1:
+        :param configuration_2:
+        :return:
+        """
+        configurations = (configuration_1, configuration_2)
+        comparable_configurations = []
+        for configuration in configurations:
+            comparable_configurations.append(
+                set(map(lambda item: (item[0], str(item[1])),
+                        ShinobiMonitorOrm.filter_only_supported_keys(configuration).items())))
+        assert len(comparable_configurations) == 2
+        return comparable_configurations[0] == comparable_configurations[1]
 
     @staticmethod
     def validate_configuration(configuration: Dict):
@@ -100,11 +117,12 @@ class ShinobiMonitorOrm:
         response.raise_for_status()
         return {entry["mid"]: entry for entry in response.json()}
 
-    def create(self, monitor_id: str,  configuration: Dict) -> Dict:
+    def create(self, monitor_id: str,  configuration: Dict, verify: bool = True) -> Dict:
         """
         TODO
         :param monitor_id:
         :param configuration:
+        :param verify:
         :return:
         :raises MonitorAlreadyExistsError:
         """
@@ -116,44 +134,71 @@ class ShinobiMonitorOrm:
             raise ShinobiMonitorAlreadyExistsError(monitor_id)
 
         self._configure(monitor_id, configuration)
-        return self.get(monitor_id)
 
-    def modify(self, monitor_id: str, configuration: Dict) -> bool:
+        retrieved_monitor = None
+
+        def retrieve_monitor():
+            nonlocal retrieved_monitor
+            retrieved_monitor = self.get(monitor_id)
+            return retrieved_monitor is not None
+
+        if verify:
+            if not wait_and_verify(retrieve_monitor):
+                raise RuntimeError(f"Could not create monitor \"{monitor_id}\" with configuration: ${configuration}")
+            assert retrieved_monitor is not None
+        else:
+            retrieved_monitor = self.get(monitor_id)
+
+        return retrieved_monitor
+
+    def modify(self, monitor_id: str, configuration: Dict, verify: bool = True) -> bool:
         """
         TODO
 
         :param monitor_id:
         :param configuration:
+        :param verify:
         :return:
         """
         ShinobiMonitorOrm.validate_configuration(configuration)
         current_configuration = self.get(monitor_id)
-        if not self.get(monitor_id):
+        if not current_configuration:
             raise ShinobiMonitorDoesNotExistError(monitor_id)
 
-        comparable_input_configuration = set(map(lambda item: (item[0], str(item[1])), configuration.items()))
-        comparable_current_configuration = set(map(lambda item: (item[0], str(item[1])),
-                                                   ShinobiMonitorOrm.filter_only_supported_keys(current_configuration)))
-        if comparable_input_configuration == comparable_current_configuration:
+        # comparable_input_configuration = set(map(lambda item: (item[0], str(item[1])), configuration.items()))
+        # comparable_current_configuration = set(
+        #     map(lambda item: (item[0], str(item[1])),
+        #         ShinobiMonitorOrm.filter_only_supported_keys(current_configuration).items()))
+        if ShinobiMonitorOrm.is_configuration_equivalent(current_configuration, configuration):
             return False
 
-        # TODO
+        # TODO: it's unclear whether the other things need to be set (else their value may be changed)?
         # configuration = {**current_configuration, **configuration}
         self._configure(monitor_id, configuration)
+
+        if verify and not wait_and_verify(lambda: ShinobiMonitorOrm.is_configuration_equivalent(
+                self.get(monitor_id), configuration)):
+            raise RuntimeError(f"Could not change configuration of monitor \"{monitor_id}\" to: {configuration}")
+
         return True
 
-    def delete(self, monitor_id: str) -> bool:
+    def delete(self, monitor_id: str, verify: bool = True) -> bool:
         """
         TODO
         :param monitor_id:
+        :param verify:
         :return:
         """
         # Note: if we don"t do this check, Shinobi errors (and the connection hangs) if asked to remove a non-existent
         #       monitor
         if not self.get(monitor_id):
             return False
+
         response = requests.post(f"{self.base_url}/configureMonitor/{self.group_key}/{monitor_id}/delete")
         raise_if_errors(response)
+        if verify and not wait_and_verify(lambda: self.get(monitor_id) is None):
+            raise RuntimeError(f"Could not delete monitor: {monitor_id}")
+
         return True
 
     def _configure(self, monitor_id: str, configuration: Dict):